+++ src/commands/osu_scores/Simulate.ts
@@ -12,7 +12,7 @@ import { OsuCommand } from '../command';
 export class Simulate extends OsuCommand {
     declare protected params: {
         mapid: number;
-        mods: string;
+        mods: osumodcalc.types.Mod[];
         acc: number;
         combo: number;
         n300: number;
@@ -61,11 +61,15 @@ export class Simulate extends OsuCommand {
         this.params.customOD = this.setParam(this.params.customOD, ['-od', '-accuracy'], 'number', {});
         this.params.customHP = this.setParam(this.params.customHP, ['-hp', '-drain', 'health'], 'number', {});
 
-        this.params.mods = this.setParam(this.params.mods, ['-mods'], 'string', {});
-
         const tmod = this.setParamMods();
-        if (tmod) {
-            this.params.mods = tmod.mods.join('');
+        if (tmod.mods) {
+            this.params.mods = tmod.mods;
+        } else {
+            const temp = this.setParam(this.params.mods, ['-mods'], 'string', {});
+            const mods = osumodcalc.mod.fromString(temp?.toUpperCase() ?? '');
+            if (mods && mods.length > 0) {
+                this.params.mods = mods;
+            }
         }
         {
             const mapTemp = this.setParamMap();
@@ -84,7 +88,7 @@ export class Simulate extends OsuCommand {
     async setParamsInteract() {
         const interaction = this.input.interaction as Discord.ChatInputCommandInteraction;
         this.params.mapid = interaction.options.getInteger('id');
-        this.params.mods = interaction.options.getString('mods');
+        this.params.mods = osumodcalc.mod.fromString(interaction.options.getString('mods')) ?? null;
         this.params.acc = interaction.options.getNumber('accuracy');
         this.params.combo = interaction.options.getInteger('combo');
         this.params.n300 = interaction.options.getInteger('n300');
@@ -117,13 +121,13 @@ export class Simulate extends OsuCommand {
             great: this.params.n300,
             ok: this.params.n100,
             meh: this.params.n50,
-            miss: this.params.nMiss ?? 0,
+            miss: this.params.nMiss,
         };
 
         const perfs = await performance.fullPerformance(
             this.params.mapid,
             0,
-            osumodcalc.mod.fromString(this.params.mods),
+            this.params?.mods ?? [],
             this.fixAcc() / 100,
             this.params.overrideSpeed,
             scorestat,
@@ -138,11 +142,15 @@ export class Simulate extends OsuCommand {
         data.debug(perfs, this.name, this.input.message?.guildId ?? this.input.interaction?.guildId, 'ppCalc');
 
         const mapPerf = await performance.calcMap({
-            mods: osumodcalc.mod.fromString(this.params.mods),
+            mods: this.params?.mods ?? [],
             mode: 0,
-            mapid: this.params.mapid,
+            mapid: this.params.mapid, 
             clockRate: this.params.overrideSpeed,
-            mapLastUpdated: new Date(this.map.last_updated)
+            mapLastUpdated: new Date(this.map.last_updated),
+            customCS: this.params.customCS,
+            customAR: this.params.customAR,
+            customOD: this.params.customOD,
+            customHP: this.params.customHP,
         });
         this.ctn.embeds = [this.setEmbed(
             perfs, mapPerf,
@@ -183,7 +191,7 @@ export class Simulate extends OsuCommand {
                 this.params.combo = tempscore.apiData.max_combo;
             }
             if (!this.params.mods) {
-                this.params.mods = tempscore.apiData.mods.map(x => x.acronym).join('') ?? 'NM';
+                this.params.mods = tempscore.apiData.mods.map(x => x.acronym) as osumodcalc.types.Mod[] ?? [];
             }
         }
         return tempscore;
@@ -196,11 +204,11 @@ export class Simulate extends OsuCommand {
             this.params.overrideBpm = this.params.overrideSpeed * this.map.bpm;
         }
 
-        if (this.params.mods.includes('DT') || this.params.mods.includes('NC')) {
+        if (this.params?.mods?.includes('DT') || this.params?.mods?.includes('NC')) {
             this.params.overrideSpeed *= 1.5;
             this.params.overrideBpm *= 1.5;
         }
-        if (this.params.mods.includes('HT')) {
+        if (this.params?.mods?.includes('HT') || this.params?.mods?.includes('DC')) {
             this.params.overrideSpeed *= 0.75;
             this.params.overrideBpm *= 1.5;
         }
@@ -222,6 +230,7 @@ export class Simulate extends OsuCommand {
     }
 
     setEmbed(perfs: rosu.PerformanceAttributes[], mapPerf: rosu.PerformanceAttributes[], title: string, useAcc: number) {
+        const stats = this.compareStats();
         const scoreEmbed = new Discord.EmbedBuilder()
             .setTitle(`Simulated play on \n\`${title}\``)
             .setURL(`https://osu.ppy.sh/b/${this.params.mapid}`)
@@ -232,7 +241,7 @@ export class Simulate extends OsuCommand {
                     value:
                         `${(useAcc)?.toFixed(2)}% | ${this.params.nMiss ?? 0}x misses
     ${this.params.combo ?? this.map.max_combo}x/**${this.map.max_combo}**x
-    ${this.params.mods ?? 'No mods'}
+    ${this.params.mods && this.params.mods.length > 0 ? this.params.mods.join('') : 'NM'}
     \`${this.params.n300}/${this.params.n100}/${this.params.n50}/${this.params.nMiss}\`
     Speed: ${this.params.overrideSpeed ?? 1}x @ ${this.params.overrideBpm ?? this.map.bpm}BPM
     `,
@@ -242,40 +251,74 @@ export class Simulate extends OsuCommand {
                     name: 'Performance',
                     value:
                         `
-    ${perfs[0].pp?.toFixed(2)}pp | ${perfs[1].pp?.toFixed(2)}pp if ${(useAcc)?.toFixed(2)}% FC
-    SS: ${mapPerf[0].pp?.toFixed(2)}
-    99: ${mapPerf[1].pp?.toFixed(2)}
-    98: ${mapPerf[2].pp?.toFixed(2)}
-    97: ${mapPerf[3].pp?.toFixed(2)}
-    96: ${mapPerf[4].pp?.toFixed(2)}
-    95: ${mapPerf[5].pp?.toFixed(2)} 
-    `
+${perfs[0].pp?.toFixed(2)}pp | ${perfs[1].pp?.toFixed(2)}pp if ${(useAcc)?.toFixed(2)}% FC
+SS: ${mapPerf[0].pp?.toFixed(2)}
+99: ${mapPerf[1].pp?.toFixed(2)}
+98: ${mapPerf[2].pp?.toFixed(2)}
+97: ${mapPerf[3].pp?.toFixed(2)}
+96: ${mapPerf[4].pp?.toFixed(2)}
+95: ${mapPerf[5].pp?.toFixed(2)} 
+`
                 },
                 {
                     name: 'Map Details',
                     value:
                         `
-    CS${this.map.cs.toString().padEnd(5, ' ')}
-    AR${this.map.ar.toString().padEnd(5, ' ')}
-    OD${this.map.accuracy.toString().padEnd(5, ' ')}
-    HP${this.map.drain.toString().padEnd(5, ' ')}
-    ${helper.emojis.mapobjs.total_length}${calculate.secondsToTime(this.map.total_length)}
-                    `,
+CS${stats.cs}
+AR${stats.ar}
+OD${stats.od}
+HP${stats.hp}
+${helper.emojis.mapobjs.total_length}${stats.length}
+`,
                     inline: true
                 },
                 {
                     name: helper.defaults.invisbleChar,
                     value:
                         `
-    ${helper.emojis.mapobjs.circle}${this.map.count_circles}
-    ${helper.emojis.mapobjs.slider}${this.map.count_sliders}
-    ${helper.emojis.mapobjs.spinner}${this.map.count_spinners}
-    ${helper.emojis.mapobjs.bpm}${this.map.bpm}
-    ${helper.emojis.mapobjs.star}${(perfs[0]?.difficulty?.stars ?? this.map.difficulty_rating)?.toFixed(2)}
-                    `,
+${helper.emojis.mapobjs.circle}${this.map.count_circles}
+${helper.emojis.mapobjs.slider}${this.map.count_sliders}
+${helper.emojis.mapobjs.spinner}${this.map.count_spinners}
+${helper.emojis.mapobjs.bpm}${this.map.bpm}
+${helper.emojis.mapobjs.star}${(perfs[0]?.difficulty?.stars ?? this.map.difficulty_rating)?.toFixed(2)}
+`,
                     inline: true
                 },
             ]);
         return scoreEmbed;
+    }
+    mapDetails() {
+        return osumodcalc.stats.modded({
+            cs: this.params.customCS ?? this.map.cs,
+            ar: this.params.customAR ?? this.map.ar,
+            od: this.params.customOD ?? this.map.accuracy,
+            hp: this.params.customHP ?? this.map.drain,
+            bpm: this.map.bpm,
+            songLength: this.map.total_length
+        }, this.params.mods ?? [], this.params.overrideSpeed ?? 1);
+    }
+    compareStats() {
+        const stats = this.mapDetails();
+        const cs = stats.cs == this.map.cs ?
+            stats.cs + '' :
+            this.map.cs + '=>' + stats.cs;
+        const ar = stats.ar == this.map.ar ?
+            stats.ar + '' :
+            this.map.ar + '=>' + stats.ar;
+        const od = stats.od == this.map.accuracy ?
+            stats.od + '' :
+            this.map.accuracy + '=>' + stats.od;
+        const hp = stats.hp == this.map.drain ?
+            stats.hp + '' :
+            this.map.drain + '=>' + stats.hp;
+        const bpm = stats.hp == this.map.bpm ?
+            stats.bpm + '' :
+            this.map.bpm + '=>' + stats.bpm;
+        const length = stats.songLength == this.map.total_length ?
+            calculate.secondsToTime(stats.songLength) :
+            calculate.secondsToTime(this.map.total_length) + '=>' + calculate.secondsToTime(stats.songLength);
+        return {
+            cs, ar, od, hp, bpm, length
+        };
     }
 }